third_party/blink/web_tests/flag-specific/highdpi/accessibility/img-srcset-1x-expected.txt

This is a testharness.js-based test.
FAIL Test that img.url retrieves the loaded image when the img uses srcset assert_equals: expected "red.png" but got "green.png"
Harness: the test ran to completion.
